requirements.txt
numpy
torch
pyalsaaudio; sys_platform == 'linux'
pyaudio; sys_platform != 'linux'
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd

